eval/video/video_requirements.txt
absl-py==2.3.1
accelerate==1.9.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
av==15.0.0
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
bitsandbytes==0.45.5
black==25.1.0
blinker==1.9.0
boto3==1.40.1
botocore==1.40.1
braceexpand==0.1.7
Brotli==1.1.0
certifi==2025.7.14
chainlit==2.6.7
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cpm-kernels==1.0.11
cycler==0.12.1
Cython==3.1.2
dataclasses-json==0.6.7
datasets==4.0.0
decord==0.6.0
deepspeed==0.17.4
Deprecated==1.2.18
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@18f69583391e5040043ca4f4bebd2c60f0ebfde0
diffusers==0.34.0
dill==0.3.8
distro==1.9.0
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
facexlib==0.3.0
fairscale==0.4.13
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
filetype==1.2.0
filterpy==1.4.5
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
gradio==5.41.1
gradio_client==1.11.0
greenlet==3.2.3
groovy==0.1.2
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.5
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgaug==0.4.0
importlib_metadata==8.7.0
inflection==0.5.1
iopath==0.1.9
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Lazify==0.4.0
lazy_loader==0.4
literalai==0.1.201
llvmlite==0.44.0
lmdb==1.7.3
loguru==0.7.3
lpips==0.1.4
lvis==0.5.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mcp==1.12.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==3.1.0
openai==1.99.1
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.44.1
opentelemetry-instrumentation-anthropic==0.44.1
opentelemetry-instrumentation-bedrock==0.44.1
opentelemetry-instrumentation-chromadb==0.44.1
opentelemetry-instrumentation-cohere==0.44.1
opentelemetry-instrumentation-crewai==0.44.1
opentelemetry-instrumentation-google-generativeai==0.44.1
opentelemetry-instrumentation-groq==0.44.1
opentelemetry-instrumentation-haystack==0.44.1
opentelemetry-instrumentation-lancedb==0.44.1
opentelemetry-instrumentation-langchain==0.44.1
opentelemetry-instrumentation-llamaindex==0.44.1
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.44.1
opentelemetry-instrumentation-mcp==0.44.1
opentelemetry-instrumentation-milvus==0.44.1
opentelemetry-instrumentation-mistralai==0.44.1
opentelemetry-instrumentation-ollama==0.44.1
opentelemetry-instrumentation-openai==0.44.1
opentelemetry-instrumentation-openai-agents==0.44.1
opentelemetry-instrumentation-pinecone==0.44.1
opentelemetry-instrumentation-qdrant==0.44.1
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.44.1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.44.1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.44.1
opentelemetry-instrumentation-transformers==0.44.1
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.44.1
opentelemetry-instrumentation-watsonx==0.44.1
opentelemetry-instrumentation-weaviate==0.44.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.11
opentelemetry-util-http==0.55b1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parameterized==0.9.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
portalocker==3.2.0
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyiqa==0.1.10
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytorchvideo==0.1.5
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.7
s3transfer==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.34.1
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sse-starlette==3.0.2
starlette==0.47.2
SwissArmyTransformer==0.4.12
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.10.0
timm==1.0.12
tokenizers==0.13.3
tomli==2.2.1
tomlkit==0.13.3
torch==2.1.0
torchaudio==2.6.0
torchvision==0.16.0
tqdm==4.67.1
traceloop-sdk==0.44.1
transformers==4.33.2
triton==2.1.0
trueskill==0.4.5
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vbench==0.1.5
wandb==0.21.0
watchfiles==0.20.0
wcwidth==0.2.13
webdataset==1.0.2
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.22.post7
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
